pyproject.toml
[tool.poetry]
name = "poetryenv"
version = "0.1.0"
description = ""
authors = ["1izard <dee1bdnatiaw@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
cleo = "^0.8.1"
"delegator.py" = "^0.1.1"
yaspin = "^0.16.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.7.9"
autopep8 = "^1.5.2"

[tool.poetry.scripts]
poetryenv = 'poetryenv.__main__:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
